+++ deepspeed/runtime/zero/__init__.py
@@ -14,4 +14,4 @@ from .tiling import TiledLinearReturnBias
 
 from .mics import MiCS_Init
 
-from .utils import unwrap_model_for_generation
+from .stage3 import unwrap_model_for_generation
